requirements.txt
streamlit==1.0.*
pillow==8.3.*